Move remoteservices/wit.go to wit/wit.go
